SiFive: Workaround for the CIP-1412 - Add fence after each FPU load.
Check if the instruction is a FPU load in the append_insn, then we
can also add the fence after the pseudo FPU loads (macros).

I see MIPS use INSN_CHIP_MASK to mark if it is a chip specific
instruction.  Therefore, I use INSN_FPU_LOAD to mark all possible
candidates that need to add the fence in the riscv_opcodes.  But
I can not define the INSN_FPU_LOAD macro in the riscv_insn_types
table, so we won't add fence after the fpu loads, which are writen
by the .insn directives.